third_party/WebKit/LayoutTests/fast/css-grid-layout/non-named-grid-line-get-set-expected.txt

Test that setting and getting grid-definition-columns and grid-definition-rows works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting grid-definition-columns and grid-definition-rows set through CSS
PASS window.getComputedStyle(gridWithFixedElement, '').getPropertyValue('grid-definition-columns') is "first 10px"
PASS window.getComputedStyle(gridWithFixedElement, '').getPropertyValue('grid-definition-rows') is "first 15px"
PASS window.getComputedStyle(gridWithPercentElement, '').getPropertyValue('grid-definition-columns') is "53% last"
PASS window.getComputedStyle(gridWithPercentElement, '').getPropertyValue('grid-definition-rows') is "27% last"
PASS window.getComputedStyle(gridWithAutoElement, '').getPropertyValue('grid-definition-columns') is "first auto"
PASS window.getComputedStyle(gridWithAutoElement, '').getPropertyValue('grid-definition-rows') is "auto last"
PASS window.getComputedStyle(gridWithMinMax, '').getPropertyValue('grid-definition-columns') is "first minmax(10%, 15px)"
PASS window.getComputedStyle(gridWithMinMax, '').getPropertyValue('grid-definition-rows') is "minmax(20px, 50%) last"
PASS window.getComputedStyle(gridWithFixedMultiple, '').getPropertyValue('grid-definition-columns') is "first nav 10px last"
PASS window.getComputedStyle(gridWithFixedMultiple, '').getPropertyValue('grid-definition-rows') is "first nav 15px last"
PASS window.getComputedStyle(gridWithPercentageSameStringMultipleTimes, '').getPropertyValue('grid-definition-columns') is "first nav 10% nav 15% last"
PASS window.getComputedStyle(gridWithPercentageSameStringMultipleTimes, '').getPropertyValue('grid-definition-rows') is "first nav2 25% nav2 75% last"

Test getting and setting grid-definition-columns and grid-definition-rows through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "first 18px"
PASS element.style.gridDefinitionColumns is "first 18px"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "66px last"
PASS element.style.gridDefinitionRows is "66px last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "first 55%"
PASS element.style.gridDefinitionColumns is "first 55%"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "40% last"
PASS element.style.gridDefinitionRows is "40% last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "first auto"
PASS element.style.gridDefinitionColumns is "first auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "auto last"
PASS element.style.gridDefinitionRows is "auto last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "first min-content"
PASS element.style.gridDefinitionColumns is "first min-content"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "min-content last"
PASS element.style.gridDefinitionRows is "min-content last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "first max-content"
PASS element.style.gridDefinitionColumns is "first max-content"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "max-content last"
PASS element.style.gridDefinitionRows is "max-content last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "first minmax(55%, 45px)"
PASS element.style.gridDefinitionColumns is "first minmax(55%, 45px)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "minmax(30px, 40%) last"
PASS element.style.gridDefinitionRows is "minmax(30px, 40%) last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "first minmax(220px, max-content)"
PASS element.style.gridDefinitionColumns is "first minmax(22em, max-content)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "minmax(max-content, 50px) last"
PASS element.style.gridDefinitionRows is "minmax(max-content, 5em) last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "first minmax(220px, min-content)"
PASS element.style.gridDefinitionColumns is "first minmax(22em, min-content)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "minmax(min-content, 50px) last"
PASS element.style.gridDefinitionRows is "minmax(min-content, 5em) last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "first minmax(min-content, max-content)"
PASS element.style.gridDefinitionColumns is "first minmax(min-content, max-content)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "minmax(max-content, min-content) last"
PASS element.style.gridDefinitionRows is "minmax(max-content, min-content) last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "first nav minmax(min-content, max-content) last"
PASS element.style.gridDefinitionColumns is "first nav minmax(min-content, max-content) last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "first nav minmax(max-content, min-content) last"
PASS element.style.gridDefinitionRows is "first nav minmax(max-content, min-content) last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "first nav minmax(min-content, max-content) nav auto last"
PASS element.style.gridDefinitionColumns is "first nav minmax(min-content, max-content) nav auto last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "first nav2 minmax(max-content, min-content) nav2 minmax(10px, 15px) last"
PASS element.style.gridDefinitionRows is "first nav2 minmax(max-content, min-content) nav2 minmax(10px, 15px) last"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "foo bar auto foo auto bar"
PASS element.style.gridDefinitionColumns is "foo bar auto foo auto bar"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "foo bar auto foo auto bar"
PASS element.style.gridDefinitionRows is "foo bar auto foo auto bar"

Test getting and setting invalid grid-definition-columns and grid-definition-rows through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "none"
PASS successfullyParsed is true

TEST COMPLETE
